+++ audiomate/corpus/assets/__init__.py
@@ -14,5 +14,6 @@ from .issuer import Artist  # noqa: F401
 from .label import Label  # noqa: F401
 from .label import LabelList  # noqa: F401
 
+from .container import Container  # noqa: F401
 from .features import FeatureContainer  # noqa: F401
 from .features import PartitioningFeatureIterator  # noqa: F401