+++ tests/corpus/assets/test_features.py
@@ -69,32 +69,6 @@ class TestFeatureContainer:
         with pytest.raises(ValueError):
             sample_feature_container.stats()
 
-    def test_append(self, tmpdir):
-        container = assets.FeatureContainer(os.path.join(tmpdir.strpath, 'container'))
-        container.open()
-
-        data = np.arange(100).reshape(20, 5)
-
-        container.append('utt-1', data[:8])
-        container.append('utt-1', data[8:])
-
-        res = container.get('utt-1', mem_map=False)
-
-        assert np.array_equal(data, res)
-
-        container.close()
-
-    def test_append_with_different_dimension_raises_error(self, tmpdir):
-        container = assets.FeatureContainer(os.path.join(tmpdir.strpath, 'container'))
-        container.open()
-
-        container.append('utt-1', np.arange(20).reshape(5, 2, 2))
-
-        with pytest.raises(ValueError):
-            container.append('utt-1', np.arange(42).reshape(7, 2, 3))
-
-        container.close()
-
 
 class TestPartitioningFeatureIterator(object):