+++ docs/notes/changelog.rst
@@ -21,6 +21,12 @@ Next Version
 * Added downloader (:class:`audiomate.corpus.io.GtzanDownloader`) for the
   `GTZAN Music/Speech <https://marsyasweb.appspot.com/download/data_sets/>`_.
 
+**Non-Breaking Changes**
+
+* Create container base class (:class:`audiomate.corpus.assets.Container`), that can be used to store arbitrary data
+  per utterance. The :class:`audiomate.corpus.assets.FeatureContainer` is now an extension of the container,
+  that provides functionality especially for features.
+
 **Fixes**
 
 * [`#58 <https://github.com/ynop/audiomate/issues/58>`_] Keep track of number of samples per frame and between frames.